refactor: type 파일 따로 분리
