+++ snapsplit/src/app/trip/createTrip/page.tsx
@@ -1,4 +1,4 @@
-import CreateTripPage from '@trip/createTrip/CreateTripPage';
+import CreateTripPage from '@/features/trip/createTrip/CreateTripPage';
 
 export default function CreateTrip() {
   return <CreateTripPage />;
