+++ snapsplit/src/features/trip/createTrip/steps/Step1_CountrySearch/CountryList.tsx
@@ -1,18 +1,6 @@
-type Props = {
-  countries: { countryId: number; countryName: string }[];
-  selected: string[];
-  onToggle: (countryName: string) => void;
-};
+import { CountryComponentProps, CountryListProps } from '@trip/createTrip/steps/Step1_CountrySearch/type';
 
-const Country = ({
-  countryName,
-  isSelected,
-  onClick,
-}: {
-  countryName: string;
-  isSelected: boolean;
-  onClick: () => void;
-}) => {
+const CountryComponent = ({ countryName, isSelected, onClick }: CountryComponentProps) => {
   return (
     <div className="flex justify-between items-center">
       <p className="text-body-2">{countryName}</p>
@@ -27,7 +15,7 @@ const Country = ({
   );
 };
 
-const CountryList = ({ countries, selected, onToggle }: Props) => {
+const CountryList = ({ countries, selected, onToggle }: CountryListProps) => {
   return (
     <div
       className="space-y-5 overflow-y-auto pb-5 scrollbar-hide scrollbar-hide::-webkit-scrollbar"
@@ -35,7 +23,7 @@ const CountryList = ({ countries, selected, onToggle }: Props) => {
     >
       {/* 고정 높이 추후 조정하기 */}
       {countries.map(({ countryName, countryId }) => (
-        <Country
+        <CountryComponent
           key={countryId}
           countryName={countryName}
           isSelected={selected.includes(countryName)}
