+++ snapsplit/src/features/trip/createTrip/steps/Step2_SelectDate/type.ts
@@ -1,0 +1,1 @@
+export type SelectDateSectionProps = { onNext: () => void };
