+++ snapsplit/src/features/trip/createTrip/steps/Step3_AddMember/type.ts
@@ -1,0 +1,1 @@
+export type AddMemberSectionProps = { onNext: () => void };
