+++ snapsplit/src/features/trip/createTrip/steps/Step4_InputTripName/type.ts
@@ -1,0 +1,1 @@
+export type InputTripNameProps = { onNext: () => void };
