+++ snapsplit/src/shared/types/ui.ts
@@ -1,0 +1,12 @@
+// Search Bar
+export enum SearchBarType {
+    WithIcon = 'WITH_ICON',
+    WithoutIcon = 'WITHOUT_ICON',
+  }
+
+export type SearchBarProps = {
+    placeholder?: string;
+    value?: string;
+    onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
+    type?: SearchBarType;
+};
